fix(StoreAddress): convert Latitude and Longitude to decimal

+++ Models/Entities/StoreAddress.cs
@@ -8,9 +8,9 @@
 
         public required string Street { get; set; }
 
-        public required double Latitude { get; set; }
+        public required decimal Latitude { get; set; }
 
-        public required double Longitude { get; set; }
+        public required decimal Longitude { get; set; }
 
         public string? PostalCode { get; set; }
